Move SQL dialog
