+++ bundles/geocatalog-tree/src/main/java/org/orbisgis/geocatalogtree/impl/jobs/DropIndex.java
@@ -45,7 +45,7 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.List;
-import org.orbisgis.geocatalogtree.impl.SQLMessageDialog;
+import org.orbisgis.sif.components.SQLMessageDialog;
 
 /**
  * Drop index job
